requirements.txt
flask
scikit-learn
torch
torchvision
opencv-python
pillow
numpy
pandas
spacy
requests
